expanding the selection and related tools

+++ src/StackedHistTool.h
@@ -42,7 +42,7 @@ class StackedHistTool{
   THStack *stack;
   TH1F *hists[25];
 
-  std::vector<int> eventCount;  // to keep track of number of events of each classification 
+  std::vector<double> eventCount;  // to keep track of number of events of each classification 
 
   double invalid_total_x;
   double invalid_total_y;
